Plot EOW in Waterbodies - fix chart in popup to use new data format
