+++ src/app/eow-data-piechart.ts
@@ -10,6 +10,7 @@ import GeoJSON from 'ol/format/GeoJSON';
 import {EOWMap} from './eow-map';
 import {PieChart} from './pie-chart';
 import {Layers} from './layers';
+import {EowDataStruct} from './eow-data-struct';
 
 const theClass = `EOWDataPieChart`;
 const htmlElementId = 'waterbody';
@@ -96,7 +97,7 @@ export default class EOWDataPieChart {
     const validData = eowDataInWaterbodies.map(e => e.eowData).filter(f => f !== null);
     if (validData.length > 0 && point[0] && point[1] && !isNaN(point[0]) && !isNaN(point[1])) {
       this.log.info(theClass, `Draw pieChart at ${point[0]}, ${point[1]})}`);
-      const preparedChartData = this.prepareChartData(validData);
+      const preparedChartData = EowDataStruct.prepareChartData(validData);
       const el = this.htmlDocument.createElement('div');
       // const img = this.htmlDocument.createElement('img');
       const id = 'pieChart-' + Math.floor(Math.random() * 100000);
@@ -104,7 +105,7 @@ export default class EOWDataPieChart {
       // img.src = 'https://www.gravatar.com/avatar/0dbc9574f3382f14a5f4c38a0aec4286?s=20';
       // el.appendChild(img);
       this.htmlDocument.getElementById(htmlElementId).appendChild(el);
-      this.pieChart.drawD3(preparedChartData, id, map.getView().getZoom() * LOG2, point);
+      this.pieChart.drawD3(preparedChartData, id, map.getView().getZoom() * LOG2);
       const pieChartMap = new Overlay({
         element: el,
         position: point,
@@ -115,68 +116,12 @@ export default class EOWDataPieChart {
       map.addOverlay(pieChartMap);
       map.on('moveend', (evt) => {
         // force a redraw so as to change size if zoom in / out
-        this.pieChart.drawD3(preparedChartData, id, map.getView().getZoom() * LOG2, point);
+        this.pieChart.drawD3(preparedChartData, id, map.getView().getZoom() * LOG2);
       });
       this.drawDebugLines(point, preparedChartData, waterBodyIndex);
     } else {
       this.log.info(theClass, `NOT Drawing pieChart at "${point[0]}", "${point[1]}")}`);
     }
-  }
-
-  /**
-   * Return Aggregated FU data as an array of objects:
-   * [
-   *  {
-   *    name: <FU Value>,
-   *    y: {
-   *       <fu value> : {count: number of that <fu value>, points: the map geo points that have those values>}
-   *    }
-   *  },
-   *  ...
-   * ]
-   * @param features - the EOWdata that is all located in the same waterbody
-   */
-  private prepareChartData(features): any {
-    const aggregateFUValues = (fuValuesInFeatures) => {
-      const eowDataReducer = (acc, currentValue) => {
-        if (currentValue.values_ && currentValue.values_.fu_value) {
-          if (acc.hasOwnProperty(currentValue.values_.fu_value)) {
-            ++acc[currentValue.values_.fu_value].count;
-            acc[currentValue.values_.fu_value].points.push(currentValue.getGeometry().getCoordinates());
-          } else {
-            acc[currentValue.values_.fu_value] = {
-              count: 1,
-              points: [currentValue.getGeometry().getCoordinates()]
-            };
-          }
-        }
-        return acc;
-      };
-      return features.reduce(eowDataReducer, {});
-    };
-    // Add zeros for all the other FUs since the colours in the pie charts are from the ordinal number of the data, NOT the value
-    // of it's "name" attribute
-    // TODO - i don't believe this is necessary, or it should be renamed 'objectToArray'
-    const setMissingFUsToZero = (fUValuesObj) => {
-      return Object.keys(fUValuesObj).map(i => {
-        return parseInt(i, 10);
-      });
-    };
-    const arrayToObject = (array) =>
-      array.reduce((obj, item) => {
-        obj[item] = item;
-        return obj;
-      }, {});
-
-    const eowDataFUValues = aggregateFUValues(features);
-    const arrayFUValues = setMissingFUsToZero(eowDataFUValues);
-    const arrayFUValuesObj = arrayToObject(arrayFUValues);
-
-    const eowData = Object.keys(arrayFUValuesObj).map(k => {
-      return {name: k, y: eowDataFUValues[k]};
-    });
-    this.log.verbose(theClass, `EOWData: ${JSON.stringify(eowData)}`);
-    return eowData;
   }
 
   /**
